Remove icon re-registered warning from all 3 composites and storybook landing page (#2506)
* remove warning from composite

* Change files

* backspace renamed to dialpadbackspace

* fix storybook

* backspace renamed to dialpadbackspace
