+++ change/@internal-react-components-17350fae-92fc-4220-811e-ca9e5873f925.json
@@ -1,0 +1,7 @@
+{
+  "type": "patch",
+  "comment": "setting icon definition to undefined if icon is already defined by fluent",
+  "packageName": "@internal/react-components",
+  "email": "carolinecao@microsoft.com",
+  "dependentChangeType": "patch"
+}
