+++ change/@internal-react-composites-a264c3d5-cdfc-4887-92b6-baf6fe3b7d75.json
@@ -1,0 +1,7 @@
+{
+  "type": "patch",
+  "comment": "setting icon definition to undefined if icon is already defined by fluent",
+  "packageName": "@internal/react-composites",
+  "email": "carolinecao@microsoft.com",
+  "dependentChangeType": "patch"
+}
