+++ change/@internal-storybook-2e8ec229-8cbf-4ab3-98f9-66f0e871777a.json
@@ -1,0 +1,7 @@
+{
+  "type": "patch",
+  "comment": "Move icon register code to manager to avoid re-register warning on load",
+  "packageName": "@internal/storybook",
+  "email": "carolinecao@microsoft.com",
+  "dependentChangeType": "patch"
+}
