+++ packages/react-components/src/theming/icons.tsx
@@ -42,11 +42,10 @@ import {
   MicProhibited16Filled,
   VideoProhibited16Filled
 } from '@fluentui/react-icons';
-/* @conditional-compile-remove(call-readiness) */
-import { Important20Filled } from '@fluentui/react-icons';
-
 /* @conditional-compile-remove(dialpad) */ /* @conditional-compile-remove(PSTN-calls) */
 import { Backspace20Regular } from '@fluentui/react-icons';
+/* @conditional-compile-remove(call-readiness) */
+import { Important20Filled } from '@fluentui/react-icons';
 
 /* @conditional-compile-remove(call-readiness) */
 import { Sparkle20Filled, VideoProhibited20Filled, MicProhibited20Filled } from '@fluentui/react-icons';
@@ -247,7 +246,7 @@ export const DEFAULT_COMPONENT_ICONS = {
   SendBoxSendHovered: <Send20Filled />,
   VideoTileMicOff: <MicOff16Filled />,
   /* @conditional-compile-remove(dialpad) */ /* @conditional-compile-remove(PSTN-calls) */
-  BackSpace: <Backspace20Regular />,
+  DialpadBackspace: <Backspace20Regular />,
   /* @conditional-compile-remove(call-readiness) */
   DomainPermissionsSparkle: <DomainPermissionSparkle20Filled />,
   /* @conditional-compile-remove(call-readiness) */
