+++ packages/storybook/.storybook/preview.tsx
@@ -1,9 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { DEFAULT_COMPONENT_ICONS, FluentThemeProvider, LocalizationProvider } from '@azure/communication-react';
-import { initializeIcons, loadTheme, registerIcons } from '@fluentui/react';
-import { HandRight20Regular, Open20Regular, Record20Regular, People20Regular } from '@fluentui/react-icons';
+import { FluentThemeProvider, LocalizationProvider } from '@azure/communication-react';
 import { Anchor, DocsContainer } from '@storybook/addon-docs/blocks';
 import React from 'react';
 import {
@@ -16,21 +14,6 @@ import {
 import { THEMES } from '../stories/themes';
 import { LOCALES } from '../stories/locales'
 import { TOC } from './TOC';
-import { initializeFileTypeIcons } from '@fluentui/react-file-type-icons';
-
-// icons used in stories that are not part of the default component icons
-const storyIcons = {
-  Open: <Open20Regular />,
-  Participants: <People20Regular />,
-  Record: <Record20Regular />,
-  RightHand: <HandRight20Regular />
-};
-
-// Removing `loadTheme({})` causes storybook declaration exception.
-loadTheme({});
-initializeIcons();
-initializeFileTypeIcons();
-registerIcons({ icons: {...DEFAULT_COMPONENT_ICONS, ...storyIcons} });
 
 export const parameters = {
   layout: 'fullscreen',
