+++ packages/storybook/stories/QuickStarts/snippets/App.snippet.tsx
@@ -2,12 +2,12 @@ import { FluentThemeProvider, DEFAULT_COMPONENT_ICONS } from '@azure/communicati
 import { initializeIcons, registerIcons } from '@fluentui/react';
 import React from 'react';
 
-function App(): JSX.Element {
-  // If you don't want to provide custom icons, you can register the default ones included with the library.
-  // This will ensure that all the icons are rendered correctly.
-  initializeIcons();
-  registerIcons({ icons: DEFAULT_COMPONENT_ICONS });
+// If you don't want to provide custom icons, you can register the default ones included with the library.
+// This will ensure that all the icons are rendered correctly.
+initializeIcons();
+registerIcons({ icons: DEFAULT_COMPONENT_ICONS });
 
+function App(): JSX.Element {
   return (
     <FluentThemeProvider>
       <h1>Hooray! You set up Fluent Theme Provider 🎉🎉🎉</h1>
